requirements.txt
# PRWhisperer — pin versions in production as needed
vllm>=0.4.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
httpx>=0.25.0
pydantic>=2.0.0
streamlit>=1.30.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-chroma>=0.1.0
chromadb>=0.4.0
sentence-transformers>=2.3.0
PyGithub>=2.1.0
rank-bm25>=0.2.2
